전남대 BE_조설빈 3주차 과제 (1, 2단계) (#327)
* feat : 2주차 코드 옮기기

* refactor : ProductAdminContrller 수정

* refactor : ProductController 수정

* refactor : WishlistController 수정

* refactor : Product 수정

* refactor : User 수정

* refactor : Wishlist 수정

* refactor : ProductRepository 수정

* refactor : WishlistRepository 수정

* refactor : ProductService 수정

* refactor : WishlistService 수정

* feat : data,schema 수정

* feat : requests.http 수정

* docs : Readme 작성

* feat : WishlistController 수정

* refactor : 데이터 스키마 자동 생성을 위한 Product, User, Wishlist 클래스 수정

* refactor : WishlistService에 @Transactonal 어노테이션 추가

* refactor : 데이터 스키마 자동 생성을 위해 application.properties, data, schema 수정

* feat : requests.http 수정

* docs : Readme 수정

+++ src/main/java/gift/service/ProductService.java
@@ -5,6 +5,7 @@ import gift.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -20,20 +21,30 @@ public class ProductService {
     }
 
     public Product getProductById(Long id) {
-        return productRepository.findById(id);
+        Optional<Product> product = productRepository.findById(id);
+        if (product.isEmpty()) {
+            throw new IllegalArgumentException("Invalid product Id:" + id);
+        }
+        return product.get();
     }
 
     public Product saveProduct(Product product) {
-        validateProductName(product.name());
+        validateProductName(product.getName());
         return productRepository.save(product);
     }
 
     public void updateProduct(Product product) {
-        validateProductName(product.name());
-        productRepository.update(product);
+        validateProductName(product.getName());
+        if (!productRepository.existsById(product.getId())) {
+            throw new IllegalArgumentException("Invalid product Id:" + product.getId());
+        }
+        productRepository.save(product);
     }
 
     public void deleteProduct(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new IllegalArgumentException("Invalid product Id:" + id);
+        }
         productRepository.deleteById(id);
     }
